feat: improve student methods security

+++ src/student/student.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from '../database/prisma.service';
 import { StudentEntity } from './student.entity';
 import { StudentCreateDto, StudentUpdateDto } from './dto';
@@ -30,7 +30,21 @@ export class StudentService {
     });
   }
 
-  async update(id: string, data: StudentUpdateDto): Promise<StudentEntity> {
+  async update(
+    organizationId: string,
+    id: string,
+    data: StudentUpdateDto,
+  ): Promise<StudentEntity> {
+    const student = await this.prisma.student.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (student.organizationId !== organizationId) {
+      throw new UnauthorizedException();
+    }
+
     return this.prisma.student.update({
       where: {
         id,
@@ -39,7 +53,17 @@ export class StudentService {
     });
   }
 
-  async delete(id: string): Promise<StudentEntity> {
+  async delete(organizationId: string, id: string): Promise<StudentEntity> {
+    const student = await this.prisma.student.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (student.organizationId !== organizationId) {
+      throw new UnauthorizedException();
+    }
+
     return this.prisma.student.delete({
       where: {
         id,
